Configuración de producción

+++ Backend/app.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 
-from config.database import Base, engine
+from config.database import Base, production, engine
 from routes.equipments import equipments
 from routes.buildings import buildings
 from routes.units import units
@@ -25,12 +25,15 @@ from routes.locations import locations
 from routes.sub_locations import sub_locations
 from routes.Suppliers_supplies import suppliers_supplies
 
+
 def create_tables():
-    Base.metadata.drop_all(bind=engine)
+    if not production:
+        Base.metadata.drop_all(bind=engine)
     Base.metadata.create_all(bind=engine)
 
 
-#create_tables()
+create_tables()
+
 app = FastAPI()
 app.mount("/images", StaticFiles(directory="images"), name="images")
 app.add_middleware(
